refactor(settings): move the secret_manager in the config dir
